+++ src/main/java/kitchenpos/order/domain/OrderRepository.java
@@ -1,8 +1,11 @@
-package kitchenpos.domain;
+package kitchenpos.order.domain;
 
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderStatus;
+import kitchenpos.order.domain.OrderTable;
 
 public interface OrderRepository {
     Order save(Order order);
