+++ src/main/java/kitchenpos/order/infra/JpaOrderTableRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.domain;
+package kitchenpos.order.infra;
 
+import kitchenpos.order.domain.OrderTable;
+import kitchenpos.order.domain.OrderTableRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
