+++ src/test/java/kitchenpos/Fixtures.java
@@ -9,6 +9,11 @@ import java.util.Random;
 import java.util.UUID;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menugroup.domain.MenuGroup;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderLineItem;
+import kitchenpos.order.domain.OrderStatus;
+import kitchenpos.order.domain.OrderTable;
+import kitchenpos.order.domain.OrderType;
 import kitchenpos.product.domain.Product;
 
 public class Fixtures {
